we can change workspaces and move tasks inside the group

+++ db/seeds.rb
@@ -7,13 +7,12 @@
 #   Mayor.create(name: 'Emanuel', city: cities.first)
 
 if Rails.env == 'development'
-  User.where.not(:email => 'neck.varentsov@gmail.com').delete_all
-  Group.delete_all
-  Workspace.delete_all
+  User.where.not(:email => ['neck.varentsov@gmail.com', 'neck@varentsov.com']).delete_all
+  Group.where.not(:account_id => nil).delete_all
   Recipient.delete_all
   Message.delete_all
   for i in 0...100
-    User.create!(email: "user#{i}@gmail.com", password: '123123', first_name: i.to_s, last_name: i.to_s)
+    User.create!(email: "user#{i}@gmail.com", password: '123123', first_name: "User#{i}", last_name: "Family#{i}")
     puts i
   end
 end
